+++ examples/intl/next.config.mjs
@@ -2,6 +2,10 @@ import withRoutes from "nextjs-routes/config";
 
 const nextConfig = {
   reactStrictMode: true,
+  i18n: {
+    locales: ["en-US", "fr", "nl-NL"],
+    defaultLocale: "en-US",
+  },
 };
 
 export default withRoutes()(nextConfig);