+++ examples/intl/src/pages/index.tsx
@@ -1,8 +1,11 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 import Link from "next/link";
+import { useRouter } from "next/router";
 
 const Home: NextPage = () => {
+  const { locale } = useRouter();
+  console.log({ locale });
   return (
     <>
       <Head>
@@ -11,9 +14,7 @@ const Home: NextPage = () => {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <main>
-        <Link href={{ pathname: "/foos/[foo]", query: { foo: "Foo" } }}>
-          Foo
-        </Link>
+        <Link href={{ pathname: "/store" }}>Store</Link>
       </main>
     </>
   );